Fix VC - A4 parse error Substrate address (#1475)
* using account_id_to_string

* capture graphql deserialize error

* ChallengeCode not found

* update ts-test

* fix A6: u64 -> u32

* add 0x & test graphql

+++ tee-worker/litentry/core/assertion-build/src/a6.rs
@@ -75,8 +75,8 @@ pub fn build(
 	}
 
 	info!("sum followers: {}", sum);
-	let min: u64;
-	let max: u64;
+	let min: u32;
+	let max: u32;
 
 	match sum {
 		0 | 1 => {
@@ -101,7 +101,7 @@ pub fn build(
 		},
 		100001..=u32::MAX => {
 			min = 100000;
-			max = u64::MAX;
+			max = u32::MAX;
 		},
 	}
